requirements.txt
# Core Django
Django>=4.2,<5.0
django-environ>=0.10.0

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL support (optional)

# Forms and UI
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2

# Image handling
Pillow>=10.0.0

# Development and debugging
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# API Development
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0  # For API documentation

# Security
django-csp>=3.7  # Content Security Policy

# Static files
whitenoise>=6.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
flake8>=6.1.0
black>=23.0.0
isort>=5.12.0

# Email (for appointment notifications)
django-anymail>=10.2

# Payment processing
stripe>=6.0.0

# Utilities
python-slugify>=8.0.0
django-model-utils>=4.3.0

# Production
gunicorn>=21.2.0
